feat: 머지 + 클린 진행중

+++ Acha/Acha/Presentation/TabBar/Home/SelectMap/SelectMapViewController.swift
@@ -321,6 +321,7 @@ extension SelectMapViewController: UICollectionViewDelegate {
     }
     
     private func makeSnapshot(rankings: [Record]) {
+        guard let dataSource else { return }
         var snapshot = dataSource.snapshot()
         snapshot.deleteAllItems()
         snapshot.appendSections(["Ranking"])
